peer coding with Dylan. work in progress. Separated stamp server part away from signing node. working on designing interfaces
